[DDW-575] Moving FormattedHTMLMessageWithLink to components dir
