+++ source/renderer/app/components/widgets/FormattedHTMLMessageWithLink.js
